[UI] Agarでのメニュー構築
git-svn-id: https://xm7-sdl.unfuddle.com/svn/xm7-sdl_xm7-sdl-linux@168 a0c0af59-af12-4589-88dc-6bbe4490f711

+++ linux-sdl/agar_cfg.cpp
@@ -99,7 +99,7 @@ static BOOL LoadCfgString(char *key, char *buf, int length)
     char          *dat;
     ASSERT(key);
     dat = INI_getString(pszSection, key, "");
-    if (strlen(dat) == 0 || strlen(dat) > length) {
+    if ((strlen(dat) == 0) || (strlen(dat) > (size_t)length)) {
 	return FALSE;
     }
     strcpy(buf, dat);
